runner moved into terminal
