--- modules/script/pkg/manifest.json
@@ -1,30 +1,0 @@
-[
-    {
-        "process_name": "runner",
-        "process_wasm_path": "/runner.wasm",
-        "on_exit": "Restart",
-        "request_networking": true,
-        "request_capabilities": [
-            "terminal:terminal:sys",
-            "filesystem:distro:sys",
-            "http_server:distro:sys",
-            "http_client:distro:sys",
-            "net:distro:sys",
-            "vfs:distro:sys",
-            "kernel:distro:sys",
-            "eth:distro:sys",
-            "sqlite:distro:sys",
-            "kv:distro:sys",
-            "chess:chess:sys",
-            "kns_indexer:kns_indexer:sys",
-            {
-                "process": "vfs:distro:sys",
-                "params": {
-                    "root": true
-                }
-            }
-        ],
-        "grant_capabilities": [],
-        "public": true
-    }
-]
