--- modules/script/runner/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "runner"
-version = "0.1.0"
-edition = "2021"
-
-[profile.release]
-panic = "abort"
-opt-level = "s"
-lto = true
-
-[dependencies]
-anyhow = "1.0"
-bincode = "1.3.3"
-kinode_process_lib = { git = "ssh://git@github.com/uqbar-dao/process_lib.git", rev = "9790c0f" }
-rand = "0.8"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-wit-bindgen = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "efcc759" }
-
-[lib]
-crate-type = ["cdylib"]
-
-[package.metadata.component]
-package = "kinode:process"
