+++ modules/terminal/pkg/manifest.json
@@ -6,8 +6,22 @@
         "request_networking": true,
         "request_capabilities": [
             "net:distro:sys",
+            "filesystem:distro:sys",
+            "http_server:distro:sys",
             "http_client:distro:sys",
-            "kernel:distro:sys"
+            "kernel:distro:sys",
+            "vfs:distro:sys",
+            "eth:distro:sys",
+            "sqlite:distro:sys",
+            "kv:distro:sys",
+            "chess:chess:sys",
+            "kns_indexer:kns_indexer:sys",
+            {
+                "process": "vfs:distro:sys",
+                "params": {
+                    "root": true
+                }
+            }
         ],
         "grant_capabilities": [],
         "public": true
