+++ modules/terminal/terminal/Cargo.toml
@@ -13,6 +13,7 @@ lto = true
 [dependencies]
 anyhow = "1.0"
 bincode = "1.3.3"
+rand = "0.8"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 kinode_process_lib = { git = "ssh://git@github.com/uqbar-dao/process_lib.git", tag = "v0.5.3-alpha" }
